Move PageHeader out of the book page
The library page leads with the same title-then-controls header, so the
component is no longer book-specific. Comments follow it: "tab" becomes
"page" wherever the rule holds for both, and stays "tab" where the
reasoning really is about a book tab.
