+++ frontend/src/pages/BookPage/ReadingStatistics/ReadingStatisticsPage.tsx
@@ -3,10 +3,10 @@ import { FadeInOut } from '@/components/animations/FadeInOut';
 import { Spinner } from '@/components/animations/Spinner.tsx';
 import { CardList } from '@/components/CardList.tsx';
 import { EmptyStateText } from '@/components/EmptyStateText.tsx';
+import { PageHeader } from '@/components/layout/PageHeader.tsx';
 import { PaginationControls } from '@/components/PaginationControls.tsx';
 import { SectionTitle } from '@/components/typography/SectionTitle.tsx';
 import { useBookPage } from '@/pages/BookPage/BookPageContext';
-import { PageHeader } from '@/pages/BookPage/common/PageHeader.tsx';
 import { BOOK_PAGE_LABELS } from '@/pages/BookPage/navigation/bookPageRoutes.ts';
 import { Alert, Box } from '@mui/material';
 import { useNavigate, useSearch } from '@tanstack/react-router';
